+++ src/restfw/resource_info.py
@@ -3,61 +3,9 @@
 :Authors: cykooz
 :Date: 06.12.2016
 """
-from zope.interface import implementer, provider
-
-from .errors import ValidationError
-from .interfaces import IResourceInfo, IResourceInfoFabric, ISendTestingRequest
+from .usage_examples import UsageExamples
 
 
-@provider(IResourceInfoFabric)
-@implementer(IResourceInfo)
-class ResourceInfo(object):
-
-    ValidationError = ValidationError
-    headers_for_listing = None
-    test_listing = True
-
-    def __init__(self, request):
-        """
-        :type request: pyramid.request.Request
-        """
-        self.registry = request.registry
-        self.root = request.root
-        self.request = request
-        self.resource = self.prepare_resource()
-        self.request.context = self.resource
-        self.resource_url = self.request.resource_url(self.resource)
-        self.allowed_methods = self.resource.get_allowed_methods()
-
-    def prepare_resource(self):
-        return None
-
-    def get_requests(self, send):
-        """
-        :type send: ISendTestingRequest
-        """
-        pass
-
-    def put_requests(self, send):
-        """
-        :type send: ISendTestingRequest
-        """
-        pass
-
-    def patch_requests(self, send):
-        """
-        :type send: ISendTestingRequest
-        """
-        pass
-
-    def post_requests(self, send):
-        """
-        :type send: ISendTestingRequest
-        """
-        pass
-
-    def delete_requests(self, send):
-        """
-        :type send: ISendTestingRequest
-        """
-        pass
+# bw compatibility
+class ResourceInfo(UsageExamples):
+    pass